third_party/WebKit/LayoutTests/inspector/sources/debugger/frameworks-pause-button.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script src="resources/framework.js"></script>
<script>

var func = Framework.bind(Framework.safeRun, null, Framework.empty, callback);

function callback()
{
    var x = 42; // Should pause here.
    return x;
}

function test()
{
    var frameworkRegexString = "/framework\\.js$";
    WebInspector.settings.skipStackFramesPattern.set(frameworkRegexString);

    InspectorTest.startDebuggerTest(step1, true);

    function step1()
    {
        InspectorTest.evaluateInPage("setTimeout(func, 200)", step2);
    }

    function step2()
    {
        InspectorTest.assertTrue(!WebInspector.panels.sources.paused());
        WebInspector.panels.sources.togglePause();
        InspectorTest.waitUntilPaused(step3);
    }

    function step3(callFrames)
    {
        InspectorTest.captureStackTrace(callFrames);
        InspectorTest.completeDebuggerTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that hitting on pause button will not stop inside blackboxed framework.
</p>
</body>
</html>
