third_party/WebKit/LayoutTests/inspector/sources/debugger/frameworks-pause-button-expected.txt

Tests that hitting on pause button will not stop inside blackboxed framework.

Call stack:
    0) callback (frameworks-pause-button.html:12)
  * 1) Framework.safeRun (framework.js:8)
  * 2) Framework.safeRun (framework.js:10)
  * 3) Framework_bound (framework.js:105)
